syerif202201-03-2

+++ pkk.js
@@ -134,23 +134,23 @@ const update = (request, response) => {
     //console.log(mmsi);
     let jenis_telkompel;
 
-    pool.query('SELECT count(*) as total FROM tbl_masdex_pkk where id=$1 and is_delete=false', [id], (error, results) => {
-        if (error) {
-          throw error
-        }else{
-            //console.log(results.rows);
-        }
+    // pool.query('SELECT count(*) as total FROM tbl_masdex_pkk where id=$1 and is_delete=false', [id], (error, results) => {
+    //     if (error) {
+    //       throw error
+    //     }else{
+    //         //console.log(results.rows);
+    //     }
         
-    })
+    // })
 
 
-    pool.query('SELECT SELECT tbl_masdex_telkompel.jenis_telkompel FROM tbl_masdex_pelabuhan INNER JOIN tbl_masdex_telkompel ON tbl_masdex_pelabuhan.telkompel_id = tbl_masdex_telkompel."id" WHERE tbl_masdex_pelabuhan."id" = $1',[pelabuhan_tujuan], (error,results)=> {
+    pool.query('SELECT tbl_masdex_telkompel.jenis_telkompel FROM tbl_masdex_pelabuhan INNER JOIN tbl_masdex_telkompel ON tbl_masdex_pelabuhan.telkompel_id = tbl_masdex_telkompel."id" WHERE tbl_masdex_pelabuhan."id" = $1',[pelabuhan_tujuan], (error,results)=> {
       if (error) {
         response.status(201).send(error)
       }
 
         jenis_telkompel = results.rows[0].jenis_telkompel;
-
+        console.log(jenis_telkompel);
     });
 
      pool.query('SELECT * FROM tbl_masdex_pkk where id=$1 and is_delete=false',[id] ,(error, results) => {
@@ -159,56 +159,45 @@ const update = (request, response) => {
           }
 
          doc = results.rows[0].dokumen_pkk;
+         //console.log(doc);
          var doc_path = __dirname +path.join('/dokumens/pkk/'+ doc);
-         console.log(doc_path);
-         fs.unlinkSync(doc_path);
-         console.log(doc_path);
+         //console.log(doc_path);
+         if(fs.existsSync(doc_path))
+         {
+            fs.unlinkSync(doc_path);
+            console.log(doc_path);
+         }
+ 
+          let sampleFile = request.files.dokumen;
+        //  console.log(sampleFile);
+           const now = Date.now();
+           let name = now + '_' + sampleFile['name'].replace(/\s+/g, '')
+           complete_path = base_url+'dokumens/pkk/'+name;
+           console.log(complete_path);
 
-         let sampleFile = request.files.dokumen;
-         console.log(sampleFile);
-          const now = Date.now()
-          let name = now + '_' + sampleFile['name'].replace(/\s+/g, '')
-          complete_path = base_url+'dokumens/pkk/'+name;
-          console.log(__dirname);
-          sampleFile.mv(path.join(__dirname + '/dokumens/pkk/') + name, function (err) {
-              if (err){
-                console.log(err);
-              }
+        //   //console.log(__dirname);
+           sampleFile.mv(path.join(__dirname + '/dokumens/pkk/') + name, function (err) {
+               if (err){
+                 console.log(err);
+               }
                   
           });
 
-          console.log(name);
-         const update_time = new Date;
-         pool.query('UPDATE tbl_masdex_pkk SET mmsi=$1, agen_kapal=$2, pelabuhan_asal=$3, pelabuhan_tujuan=$4, pelabuhan_selanjutnya=$5, dermaga_tujuan=$6, area_tambat_tujuan=$7,zona_labuh_tujuan=$8,jenis_muatan=$9,maksimal_draft=$10,eta=$11,etd=$12,jenis_pelayaran=$13,nama_nahkoda=$14,telepon_nahkoda=$15,dokumen_pkk=$16,updated_at=$17,url_pkk=$18,no_dokumen=$19 where id=$20'
-         , [mmsi, agen_kapal, pelabuhan_asal, pelabuhan_tujuan, pelabuhan_selanjutnya, dermaga_tujuan, area_tambat_tujuan,zona_labuh_tujuan,jenis_muatan,maksimal_draft,eta,etd,jenis_pelayaran,nama_nahkoda,telepon_nahkoda,name,update_time,complete_path,no_dokumen,id], (error, results) =>{
-           if (error) {
-             // throw error
-             //response.status(201).send(error)
-             //console.log(error);
-             if (error.code == '23505')
-             {
-                 //console.log("\n ERROR! \n Individual with name: " + body.fname + " " + body.lname + " and phone #: " + body.phone + " is a duplicate member. \n");
-                 response.status(400).send('Duplicate data')
-                 return;
-             }
-           }else
-           {
-               response.status(200).send({success:true,data:'data pkk berhasil diperbarui'})
-           }
+        //  console.log(name);
+            const update_time = new Date;
+            pool.query('UPDATE tbl_masdex_pkk SET mmsi=$1, agen_kapal=$2, pelabuhan_asal=$3, pelabuhan_tujuan=$4, pelabuhan_selanjutnya=$5, dermaga_tujuan=$6, area_tambat_tujuan=$7,zona_labuh_tujuan=$8,jenis_muatan=$9,maksimal_draft=$10,eta=$11,etd=$12,jenis_pelayaran=$13,nama_nahkoda=$14,telepon_nahkoda=$15,dokumen_pkk=$16,updated_at=$17,url_pkk=$18,no_dokumen=$19 where id=$20'
+          , [mmsi, agen_kapal, pelabuhan_asal, pelabuhan_tujuan, pelabuhan_selanjutnya, dermaga_tujuan, area_tambat_tujuan,zona_labuh_tujuan,jenis_muatan,maksimal_draft,eta,etd,jenis_pelayaran,nama_nahkoda,telepon_nahkoda,name,update_time,complete_path,no_dokumen,id], (error, results) =>{
+            if (error) {
+               throw error
+            }else
+            {
+                response.status(200).send({success:true,data:'data pkk berhasil diperbarui'})
+            }
      
-         })
-
-
-
+          });
 
         });
 
-        
-
-   
-
-   
-    
 }
 
 const delete_ = (request, response) => {
@@ -247,9 +236,6 @@ const delete_ = (request, response) => {
            }
      
          })
-
-
-
 
         });
